-Began working on feature that always showed defunct attractions if the park is also defunct, see the progress by searching for allAttractionsAreDefunt in AttractionsViewController -Defunct and Progress labels do not appear on the Attractions View until the attraction data has downloaded -Pulled the attraction table view up a bit, so the attraction tally box isn't pressed right up against the bottom -Added the compose icon for modify attraction
I am proposing a bit of a redesign of the background of the Attraction and Add Park views. Give it a few days to get used to it and we can see which one we want to use. For the new design to work, The check and increment images would ave to be redesigned to the lighter green used in the Parks view. Justin, can you quickly updated those images and try them out?
-Also added the darkenBackground to Park View so the background darkens when the add park or attraction views are on top